Minor updates/fixes
Updated the back button for posts to go back to the previous route instead of defaulting to manage post page. Updated positioning of 'see all' link for posts in profile page. Added a p tag to show no comments received. Added a new class method for posts to use in pagination.